refactor: 💡 clean up registration scripts and ajax call
