+++ src/components/apiWrapper/ts/ajaxCallRegistration.ts
@@ -1,89 +1,13 @@
 import { config } from '../../../resources/ts/config';
 import { initRegistrationOverlay } from '../../overlay/ts/handleRegistrationOverlay';
-import { warningLabelForTranslatableAndParentId } from '../../registrationFormular/ts/warningLabels';
+import { warningLabelForTranslatableAndParentId } from '../../registration/ts/warningLabels';
 import { generateCsrfToken } from '../../../resources/ts/helpers/generateCsrfToken';
 import { autoLogin } from '../../registration/ts/autoLogin';
 import {
 	getErrorCaseForStatus,
 	redirectToErrorPage
 } from '../../errorPage/ts/errorHandling';
-import { getConsultingTypeFromRegistration } from '../../../resources/ts/helpers/resorts';
-import { hasPostcodeToBeExtended } from '../../postcodeSuggestion/ts/postcodeSuggestionHelper';
 import { removeAllCookies } from '../../sessionCookie/ts/accessSessionCookie';
-import { extendPostcodeToBeValid } from '../../registration/ts/registrationHelper';
-
-// export const initRegistrationCall = (e: Event) => {
-// 	const registrationData = getRegistrationDataObject();
-// 	postRegistration(config.endpoints.registerAsker, registrationData);
-// };
-
-// const getRegistrationDataObject = () => {
-// 	const postcode = getValidPostcode();
-// 	const email = document.getElementById('email') as HTMLInputElement;
-// 	const addictiveDrugs = getAddictiveDrugs();
-// 	const relation = getSelectedRadioButton('relation');
-// 	const age = document.getElementById('ageSelect') as HTMLInputElement;
-// 	const state = document.getElementById('stateSelect') as HTMLInputElement;
-// 	const gender = getSelectedRadioButton('gender');
-// 	const consultingType = getConsultingTypeFromRegistration().toString();
-
-// 	const registrationData = {
-// 		username: (document.getElementById('username') as HTMLInputElement)
-// 			.value,
-// 		postcode: postcode,
-// 		password: encodeURIComponent(
-// 			(document.getElementById('passwordInput') as HTMLInputElement).value
-// 		),
-// 		...(email && email.value != '' ? { email: email.value } : null),
-// 		...(addictiveDrugs ? { addictiveDrugs: addictiveDrugs } : null),
-// 		...(relation ? { relation: relation } : null),
-// 		...(age && age.value != '' ? { age: age.value } : null),
-// 		...(state && state.value != '' ? { state: state.value } : null),
-// 		...(gender ? { gender: gender } : null),
-// 		termsAccepted: (document.getElementById(
-// 			'termsAccepted'
-// 		) as HTMLInputElement).checked.toString(),
-// 		consultingType: consultingType,
-// 		agencyId: document.getElementById('postcode').dataset.agencyId
-// 	};
-// 	return registrationData;
-// };
-
-// const getValidPostcode = () => {
-// 	let validPostcode;
-// 	const postcode = (document.getElementById('postcode') as HTMLInputElement)
-// 		.value;
-// 	if (!parseInt(postcode)) {
-// 		let inputVal = document.getElementById('postcode').dataset.inputVal;
-// 		if (hasPostcodeToBeExtended(inputVal.length)) {
-// 			inputVal = extendPostcodeToBeValid(inputVal);
-// 		}
-// 		validPostcode = inputVal;
-// 	} else {
-// 		validPostcode = postcode;
-// 	}
-// 	return validPostcode;
-// };
-
-// const getAddictiveDrugs = () => {
-// 	const inputElements = document.querySelectorAll('.multiselectItem input');
-// 	let addictiveDrugs = [];
-// 	Array.from(inputElements).forEach((inputElement) => {
-// 		const input = inputElement as HTMLInputElement;
-// 		input.checked ? addictiveDrugs.push(input.value) : '';
-// 	});
-// 	return addictiveDrugs.join(',');
-// };
-
-// const getSelectedRadioButton = (collectionName: string) => {
-// 	const radioElements = document.getElementsByName(collectionName);
-// 	let radioValue = '';
-// 	Array.from(radioElements).forEach((radioElement) => {
-// 		const radioItem = radioElement as HTMLInputElement;
-// 		radioItem.checked ? (radioValue = radioItem.value) : '';
-// 	});
-// 	return radioValue;
-// };
 
 export const postRegistration = (url: string, data: {}) => {
 	let isRequestInProgress = false;
