+++ src/components/registration/ts/warningLabels.ts
